requirements.txt
opencv-python
python-twitter
wget
numpy
